Make scheduler remotely callable.
There is a problem using WP Cron on remote server in that
the Cron schedules are being reported on a past date, instead
of the present.

One solution to this Cron problem is to just make the scheduler
function directly callable by an external service which can call
this function at regular intervals ( instead of relying on
site traffic ). This can also be configured on a server level.